Fix for overloaded 'abs(uint32_t)' is ambiguous

+++ includes/matrix_lamp/effects.h
@@ -12414,11 +12414,11 @@ void Serpentine() {
 
     // change color --------
     CRGB col1 = CHSV(ms / 29 + y * 256 / (HEIGHT - 1) + 128, 255, 255 - (HEIGHT - y) * BR_INTERWAL);
-    CRGB col2 = CHSV(ms / 29 + y * 256 / (HEIGHT - 1), 255, 255 - (HEIGHT - y) * BR_INTERWAL);
+    CRGB col2 = CHSV(ms / 29 + y * 256 / (HEIGHT - 1),       255, 255 - (HEIGHT - y) * BR_INTERWAL);
     // CRGB col3 = CHSV(ms / 29 + y * 256 / (HEIGHT - 1) + step, 255, 255 - (HEIGHT - y) * BR_INTERWAL - fade);
 
-    wu_pixel( (uint32_t)(x1 + hue * DELTA), (uint32_t)(yy - PADDING * (255 - hue)), &col1);
-    wu_pixel( abs((uint32_t)((WIDTH - 1) * 256 - (x1 + hue * DELTA))), (uint32_t)(yy - PADDING * hue), &col2);
+    wu_pixel( (uint32_t)(x1 + hue * DELTA),                                 (uint32_t)(yy - PADDING * (255 - hue)), &col1);
+    wu_pixel( (uint32_t)abs((int)((WIDTH - 1) * 256 - (x1 + hue * DELTA))), (uint32_t)(yy - PADDING * hue),         &col2);
   }
 
   step++;
